Count tasks by task-document not by task historyState (#54)
The current implementation counts the number of elements in the task
history - or the number of times tasks have change state. So if a task
moves from Draft (in the future) to Ready (now showing) - the current
dashboards will show this as two separate tasks (one Draft and one
Ready).

The most important thing we do in App Monitoring's tasks tab is to
monitor for task explosions and finding erroneous tasks. Task explosions
happen when too many task documents are generated - not when tasks
frequently change state.

This change makes it so there is one task event monitored on the
documents `authoredOn` date. The state of the task is displayed as that
task's last known state.

#50